created conftests files and cleaned code
