+++ fibonacci/tests/test_performance.py
@@ -1,9 +1,7 @@
-from time import sleep
-
 import pytest
 
 from fibonacci.dynamic import fibonacci_dynamic_v2
-from conftest import track_performance
+from fibonacci.conftest import track_performance
 
 
 @pytest.mark.performance
